- 스냅샷 모듈 적용 todo) 스냅샷 실행시 워닝뜨고, 이미지 제대로 캡쳐 안되는것 수정
